Tests/LibWeb/Text/input/navigation/ui-process-session-history-same-document-back.html

<!doctype html>
<script src="../include.js"></script>
<script>
    function queryFromURL(url) {
        return new URL(url).search || "(none)";
    }

    function relevantEntryQueries(entries, queries) {
        return entries
            .filter(entry => queries.includes(queryFromURL(entry.url)))
            .map(entry => `${queryFromURL(entry.url)}${entry.current ? "*" : ""}`)
            .join(",");
    }

    function currentUIHistory() {
        return JSON.parse(internals.dumpUIProcessSessionHistory());
    }

    function waitForPopState() {
        return new Promise(resolve => addEventListener("popstate", resolve, { once: true }));
    }

    function dump(label, uiHistory) {
        println(`${label}: location=${queryFromURL(location.href)}`);
        println(`${label}: ui-current=${queryFromURL(uiHistory.currentURL)}`);
        println(`${label}: ui-relevant-entries=${relevantEntryQueries(uiHistory.entries, ["?same-document-back-replaced", "?same-document-back-pushed"])}`);
        println(`${label}: ui-webcontent-known-relevant-entries=${relevantEntryQueries(uiHistory.webContentKnownEntries, ["?same-document-back-replaced", "?same-document-back-pushed"])}`);
        println(`${label}: ui-forward-button=${uiHistory.forwardButtonEnabled}`);
        println(`${label}: ui-uses-ui-step-coordinates=${uiHistory.webContentUsesUIStepCoordinates}`);
    }

    asyncTest(async done => {
        history.replaceState({ replaced: true }, "", "?same-document-back-replaced");
        history.pushState({ pushed: true }, "", "?same-document-back-pushed");

        const initialUIHistory = currentUIHistory();
        if (!internals.dumpSessionHistory().includes("ui-process-session-history-same-document-back.html?same-document-back-pushed (current)")
            || queryFromURL(initialUIHistory.currentURL) !== "?same-document-back-pushed"
            || relevantEntryQueries(initialUIHistory.entries, ["?same-document-back-replaced", "?same-document-back-pushed"]) !== "?same-document-back-replaced,?same-document-back-pushed*"
            || !initialUIHistory.backButtonEnabled
            || initialUIHistory.forwardButtonEnabled)
            println(`FAIL: unexpected initial UI history ${JSON.stringify(initialUIHistory)}`);

        const popStatePromise = waitForPopState();
        history.back();
        await popStatePromise;
        await internals.flushSessionHistoryTraversalQueue();

        const uiHistory = currentUIHistory();
        if (!internals.dumpSessionHistory().includes("ui-process-session-history-same-document-back.html?same-document-back-replaced (current)")
            || queryFromURL(uiHistory.currentURL) !== "?same-document-back-replaced"
            || relevantEntryQueries(uiHistory.entries, ["?same-document-back-replaced", "?same-document-back-pushed"]) !== "?same-document-back-replaced*,?same-document-back-pushed"
            || !uiHistory.forwardButtonEnabled)
            println(`FAIL: unexpected UI history ${JSON.stringify(uiHistory)}`);

        dump("after-back", uiHistory);
        done();
    });
</script>
