Tests/LibWeb/Text/expected/navigation/ui-process-session-history-same-document-back.txt

after-back: location=?same-document-back-replaced
after-back: ui-current=?same-document-back-replaced
after-back: ui-relevant-entries=?same-document-back-replaced*,?same-document-back-pushed
after-back: ui-webcontent-known-relevant-entries=?same-document-back-replaced*,?same-document-back-pushed
after-back: ui-forward-button=true
after-back: ui-uses-ui-step-coordinates=true
